refactor: move button models into component's folder
